+++ src/celengine/dsodb.h
@@ -70,7 +70,7 @@ class DSODatabase
 
     static DSODatabase* read(std::istream&);
 
-    double getAverageAbsoluteMagnitude() const;
+    float getAverageAbsoluteMagnitude() const;
 
 private:
     void buildIndexes();
@@ -85,7 +85,7 @@ private:
     DSOOctree*       octreeRoot{ nullptr };
     AstroCatalog::IndexNumber nextAutoCatalogNumber{ 0xfffffffe };
 
-    double           avgAbsMag{ 0.0 };
+    float            avgAbsMag{ 0.0f };
 };
